pasiva/caparazon
modifcacion 2
-clases y metodos
